requirements.txt
pydevd-odoo
